requirements.txt
Flask
tensorflow # or the version of TensorFlow you're using
numpy
Pillow
werkzeug
